refactor: split all method and types on an specific type
